ceil(3000000.5) shifts 1 into the sign bit
  lib/msun/src/s_ceil.c:59    j = i1 + (1<<(52-j0));
  lib/msun/src/s_floor.c:60   j = i1+(1<<(52-j0));

That line is in the branch for j0 in [21,51] - j0 is the unbiased exponent,
so |x| in [2^21, 2^52) with a fractional part. At j0 == 21 the shift is
1 << 31, which is not representable in a signed int: C17 6.5.7p4,
undefined.

Nothing exotic reaches it. |x| in [2^21, 2^22) is 2097152.0 to 4194304.0,
and ceil(3000000.5) is in the middle of that.

  $ gcc -fsanitize=undefined -fno-sanitize-recover=all
  x=3000000.5  j0=21  shift=31
  runtime error: left shift of 1 by 31 places cannot be represented
                 in type 'int'

(u_int32_t)1 on the shift. That fixes the addition too: i1 is int32_t and
the sum was signed, so `i1 + (1<<30)` could overflow for a large i1 even
where the shift was fine. With an unsigned right operand the whole
expression is unsigned and wraps by definition, which is what the code
already assumes - the next line is `if(j<i1) i0+=1;`, a carry check that
only means anything under wrapping.

Bit-identical: 6,200,000 cases over every j0 in [21,51], zero mismatches,
under UBSan. CBMC after: ceil PROVED, floor PROVED.

== two the same triage rejected, and why that matters

s_significand.c came back `arithmetic overflow on signed unary minus in
-return_value_ilogb`, which looks exactly like the rint finding. It is
not. math.h:45 defines

  #define FP_ILOGB0  (-__INT_MAX)

so ilogb() returns -2147483647 and not INT_MIN, and negating it is fine.
FreeBSD chose that value for this reason. CBMC does not model ilogb, so
its return was unconstrained and INT_MIN was in range.

s_cosl, s_sinl and s_tanl came back `-n` for the same reason: n is written
by __ieee754_rem_pio2l, which is also unmodelled.

That is the fifth calibration doing its job - an extern function's return
is as unconstrained as a pointer parameter - and it is why the 165 raw
failures are bucketed before anything is called a defect:

  85  pointer/memory      needs a precondition
  43  float div-by-zero   IEEE-754 defines it and msun depends on it
  37  exported arithmetic  the candidate set, still needing triage each

371 functions PROVED, up from 322, on 1073 of 1306 translation units.

One candidate is reported and NOT fixed: nice(int incr) overflows on
prio + incr for large incr, confirmed under UBSan. Defining that changes
what an exported libc function does with an argument POSIX does not bound,
which is a decision about behaviour rather than a repair of a mistake.

+++ hbsd/src/lib/msun/src/s_floor.c
@@ -57,7 +57,7 @@ floor(double x)
 		if(i0<0) {
 		    if(j0==20) i0+=1;
 		    else {
-			j = i1+(1<<(52-j0));
+			j = i1+((u_int32_t)1<<(52-j0));
 			if(j<i1) i0 +=1 ; 	/* got a carry */
 			i1=j;
 		    }
